Assistant checkpoint: Fix array indexing errors
Assistant generated file changes:
- bot/handlers/message_handlers.py: Fix array indexing error in sample_command
- bot/utils/rasch_analysis.py: Fix boolean indexing in person ability estimation, Fix boolean indexing in standard error calculation

---

User prompt:

❌ Xatolik yuz berdi: arrays used as indices must be of integer (or boolean) type

Iltimos, qayta urinib ko'ring.
nima muammo buldi

+++ bot/handlers/message_handlers.py
@@ -95,7 +95,7 @@ async def sample_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
         
         # Perform Rasch analysis
         analyzer = RaschAnalyzer()
-        results = analyzer.fit(data)
+        results = analyzer.fit(data.astype(float))
         
         pdf_generator = PDFReportGenerator()
